# app/openapi/gensrc/src/main/kotlin/com/codingmill/bookapi/generated/data/ErrorDTO.kt
package com.codingmill.bookapi.generated.data

import java.util.Objects
import com.fasterxml.jackson.annotation.JsonProperty
import javax.validation.constraints.DecimalMax
import javax.validation.constraints.DecimalMin
import javax.validation.constraints.Email
import javax.validation.constraints.Max
import javax.validation.constraints.Min
import javax.validation.constraints.NotNull
import javax.validation.constraints.Pattern
import javax.validation.constraints.Size
import javax.validation.Valid
import io.swagger.v3.oas.annotations.media.Schema

/**
 * 
 * @param code 
 * @param message 
 */
data class ErrorDTO(

    @Schema(example = "null", required = true, description = "")
    @field:JsonProperty("code", required = true) val code: kotlin.Int,

    @Schema(example = "null", required = true, description = "")
    @field:JsonProperty("message", required = true) val message: kotlin.String
) {

}